backend/pyproject.toml
[tool.poetry]
name = "pulsehr"
version = "0.2.0"
description = "PulseHR - Enterprise HR Management Platform"
authors = ["PulseHR Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^6.0.4"
djangorestframework = "^3.17.1"
psycopg = {extras = ["binary"], version = "^3.3.3"}
python-dotenv = "^1.2.2"
reportlab = ">=4.0.0"
openpyxl = ">=3.1.0"
gunicorn = ">=22.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"